requirements.txt
jiwer==3.0.3
openai-whisper==20240930
PyYAML==6.0.1
paramiko==3.4.0
